requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
pyasn1==0.4.8
PyYAML==6.0
requests==2.28.0
rsa==4.8
telebot==0.0.4
Telethon==1.24.0
urllib3==1.26.9
